Move camera info
